requirements.txt
anthropic>=0.18.0
python-telegram-bot==13.7
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
python-telegram-bot==13.7
python-dotenv==1.0.1
anthropic>=0.18.0